week1 코드 개선 (class를 테스트를 위해 별개의 파일로 분리)
header 분석용 class 파일생성 (HttpRequestHandler.java)
response builder class 파일 생성 (HttpResponseBuilder.java)
